requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
psutil==5.9.8
pymongo==4.6.1
dnspython==2.5.0
gunicorn==21.2.0
python-dotenv==1.0.1
certifi==2024.2.2
websockets==12.0
httpx==0.27.0